requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
python-multipart
pydantic
httpx
